test: divide mpi test files into 2 files based on need of mpiexec
